Fix error while loading systemd-sysctl.service
The parameters given in /etc/sysctl.d/net-tuning.conf doesn't need to be quoted, being quoted systemd fails because it pass the argument in the '"value1 value2 value3"' format (note the single quotes that systemd put). Unquoting it systemd pass it in the format 'value1 value2 value3' that's the correct format.